refactor(schedulecard refac): broke down the schedule card component
