+++ client/src/components/Cards/ScheduleCard/CronForm/CronForm.tsx
@@ -1,15 +1,13 @@
-import React, { useEffect, useState } from 'react';
-
-import { postCrons, getCrons, patchCrons, deleteCrons } from 'services/apiCrons/apiCrons';
-import Select from 'components/Select/Select';
-import { IAddCrons, ICron, IPatchCrons } from 'types/cronsInterfaces';
-
-import { Settings } from 'assets/index';
-import styles from './CronForm.module.scss';
-import ScheduledCron from '../ScheduledCron/ScheduledCron';
+import React, { useState } from 'react';
 
 import Day from '../Days/Days';
+import Select from 'components/Select/Select';
+import { Settings } from 'assets/index';
+
 import { ICurrentWeather } from 'types/weatherInterfaces';
+import { ICron } from 'types/cronsInterfaces';
+
+import styles from './CronForm.module.scss';
 
 const daysInWeek = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
 const refArr = ['minutes', 'hours', 'weeks', 'months', 'days'];
@@ -22,21 +20,15 @@ interface Props {
   currentWeather?: ICurrentWeather;
   controllerId: string;
   controllerTopic: string;
+  handleScheduleCron: Function;
+  scheduledCrons: ICron[];
+  error: boolean;
 }
 
-const CronForm: React.FC<Props> = ({ currentWeather, controllerId, controllerTopic }) => {
+const CronForm: React.FC<Props> = ({ currentWeather, controllerId, controllerTopic, handleScheduleCron, scheduledCrons, error }) => {
   const [cron, setCron] = useState<string[]>(Array(5).fill('*'));
   const [activeDays, setActiveDays] = useState<number[]>([]);
-  const [scheduledCrons, setScheduledCrons] = useState<ICron[]>([]);
   const [duration, setDuration] = useState<number | string>(5);
-  const [error, setError] = useState<boolean>(false);
-
-  useEffect(() => {
-    getCrons().then((crons) => {
-      if (!crons) return;
-      else setScheduledCrons(crons);
-    });
-  }, []);
 
   function handleSelectDayFn(dayIdx: number, active: boolean) {
     let _activeDays;
@@ -59,49 +51,6 @@ const CronForm: React.FC<Props> = ({ currentWeather, controllerId, controllerTop
     const _cron = convertToCronArray('minutes', [minutes], _cronHours);
 
     setCron(_cron);
-  }
-
-  function handleScheduleCron() {
-    const _cronTimeString = cron.join(' ');
-
-    const addCronObj: IAddCrons = {
-      time: _cronTimeString,
-      controllerId: controllerId,
-      action: {
-        id: controllerTopic,
-        action: 'on',
-        duration: +duration
-      }
-    };
-
-    postCrons(addCronObj).then((cron) => {
-      if (!cron) {
-        setError(true);
-        return;
-      }
-
-      setScheduledCrons([...scheduledCrons, cron]);
-    });
-  }
-
-  function handlePatchCron(id: string, updateObj: IPatchCrons) {
-    patchCrons(id, updateObj);
-  }
-
-  function handleDeleteCron(id: string) {
-    deleteCrons(id).then((res) => {
-      if (!res) return;
-      if (res.affected === 1) {
-        const _scheduledCrons = scheduledCrons.filter((cron) => cron.id !== id);
-        setScheduledCrons(_scheduledCrons);
-      }
-    });
-  }
-
-  function renderScheduledCrons() {
-    return scheduledCrons.map((cron) => {
-      return <ScheduledCron key={cron.id} cron={cron} handlePatchCron={handlePatchCron} handleDeleteCron={handleDeleteCron} />;
-    });
   }
 
   const handleDuration = (label: string, value: number | string) => setDuration(value);
@@ -128,12 +77,7 @@ const CronForm: React.FC<Props> = ({ currentWeather, controllerId, controllerTop
         </div>
       </div>
 
-      <div className={`${styles.cronPanelModule} ${styles.scrollPanelModule}`}>
-        <h2>Scheduled Actions</h2>
-        {renderScheduledCrons()}
-      </div>
-
-      <button className={styles.cronScheduleButton} onClick={handleScheduleCron}>
+      <button className={styles.cronScheduleButton} onClick={() => handleScheduleCron(cron, duration)}>
         <span>Schedule</span>
       </button>
